requirements.txt
# Core Deep Learning
torch==2.1.0+cu121 -f https://download.pytorch.org/whl/cu121
torchvision==0.16.0+cu121 -f https://download.pytorch.org/whl/cu121
torchaudio==2.1.0+cu121 -f https://download.pytorch.org/whl/cu121

# MMSegmentation Ecosystem
mmcv==2.1.0 -f https://download.openmmlab.com/mmcv/dist/cu121/torch2.1.0/index.html
mmengine>=0.9.0,<0.10.0
mmsegmentation==1.2.2
openmim>=0.3.9,<0.4.0

# Data Processing & Image Manipulation
numpy>=1.23.0,<1.27.0
pillow>=9.0.0,<11.0.0
opencv-python-headless>=4.8.0,<4.10.0
scikit-image>=0.21.0,<0.24.0

# Visualization & Analysis
matplotlib>=3.7.0,<3.10.0
seaborn>=0.12.0,<0.14.0
tensorboard>=2.13.0,<2.15.0
wandb>=0.14.0,<0.16.0

# Data Processing
pandas>=2.0.0,<2.3.0
scikit-learn>=1.3.0,<1.5.0

# Utilities & Formatting
pyyaml>=6.0,<7.0
tqdm>=4.65.0,<4.67.0
terminaltables>=3.1.10,<4.0.0
prettytable>=3.8.0,<4.0.0
rich>=13.0.0,<14.0.0

# Optional: Performance Profiling & Debugging
memory-profiler>=0.61.0,<0.63.0
py-spy>=0.3.14,<0.4.0
psutil>=5.9.0,<5.10.0

# Optional: Model Export & Deployment
onnx>=1.14.0,<1.16.0
onnxruntime>=1.16.0,<1.17.0

ftfy>=6.1.1,<6.2.0
regex>=2023.6.3,<2024.0.0